Add lazy loading component to domains list
Implemented a LazyLoader component to handle infinite scroll and data fetching in the domains list. Refactored the domains list to use LazyLoader, simplifying the existing code and centralizing the lazy loading logic. Adjusted the Vite configuration for conditional HTTPS setup.
